+++ dimas/src/com/pubsub/publisher_builder.rs
@@ -51,7 +51,7 @@ where
 	pub fn new(context: Context<P>) -> Self {
 		Self {
 			context,
-			activation_state: OperationState::Standby,
+			activation_state: OperationState::Active,
 			#[cfg(feature = "unstable")]
 			allowed_destination: Locality::Any,
 			congestion_control: CongestionControl::Drop,