+++ dimas/src/com/queries/queryable_builder.rs
@@ -55,7 +55,7 @@ where
 	pub const fn new(context: Context<P>) -> Self {
 		Self {
 			context,
-			activation_state: OperationState::Standby,
+			activation_state: OperationState::Active,
 			completeness: true,
 			#[cfg(feature = "unstable")]
 			allowed_origin: Locality::Any,